refactor: move unique_rows function to dev_utils, remove pandas dependency from utils
