+++ src/aind_data_schema_models/utils.py
@@ -4,26 +4,6 @@ import re
 from pydantic import BaseModel, Field
 from typing import Union, List, Type, Any
 from typing_extensions import Annotated
-import pandas as pd
-
-
-def unique_rows(value, key):
-    """Generate a unique subset of a dataframe based on a key column.
-    
-    Parameters
-    ----------
-    data : pd.DataFrame
-      The data to filter.
-    key : str
-      The column to filter on.
-    """
-    seen = set()
-    unique_rows = []
-    for _, row in value.iterrows():
-        if row[key] not in seen:
-            seen.add(row[key])
-            unique_rows.append(row)
-    return pd.DataFrame(unique_rows)
 
 
 def to_class_name_underscored(name: str) -> str:
